+++ src/main/java/kdt/dev/ecommerce/auth/application/OAuthFactory.java
@@ -1,4 +1,4 @@
-package kdt.dev.ecommerce.auth.applicaiton;
+package kdt.dev.ecommerce.auth.application;
 
 import static kdt.dev.ecommerce.auth.exception.AuthError.*;
