+++ src/main/java/kdt/dev/ecommerce/auth/application/dto/AuthTokenResponse.java
@@ -1,4 +1,4 @@
-package kdt.dev.ecommerce.auth.applicaiton.dto;
+package kdt.dev.ecommerce.auth.application.dto;
 
 public record AuthTokenResponse(
 	String accessToken,
